refactor: move index
