+++ src/index.tsx
@@ -27,10 +27,10 @@
  *  - [Break] RangePicker go to end field, `pickerValue` will follow the start field if not controlled.
  */
 
-import type { PickerRef } from '../interface';
-import RangePicker, { type RangePickerProps } from './PickerInput/RangePicker';
-import Picker, { type PickerProps } from './PickerInput/SinglePicker';
-import PickerPanel, { type PickerPanelProps } from './PickerPanel';
+import type { PickerRef } from './interface';
+import RangePicker, { type RangePickerProps } from './NewPicker/PickerInput/RangePicker';
+import Picker, { type PickerProps } from './NewPicker/PickerInput/SinglePicker';
+import PickerPanel, { type PickerPanelProps } from './NewPicker/PickerPanel';
 
 export { Picker, RangePicker, PickerPanel };
 export type { RangePickerProps, PickerProps, PickerPanelProps, PickerRef };
